Update the location of the data to cmc_ids
Instead of trying to anticipate which chains or protocols the user may be
interested in a watchlist tab, just get all available Coinmarketcap and
Defillama protocols and chains and match up using Defillama's cmcId. Then
update the MC, FDMC, and TVL in the cmc_id tab. This way, there is no need
to handle the scenario that a protocol of interest is listed on
Coinmarketcap but not listed on Defillama.