Updates defillama_slugs for unit testing
Added checks so that the number of protocols =
cmcId + the number of protocols not matched equals +
defillama listed cmcIds that cannot be found on CMC.

Added the same check for the chains.

Also corrected an error where chains listed only on defillama
are getting written multiple times at the end of the Excel.